split identity integration tests into separate classes
