+++ backend/DevUp/tests/DevUp.Api.Tests.Integration/V1/Identity/IdentityControllerTests_Refresh.cs
@@ -12,95 +12,17 @@ using Xunit;
 
 namespace DevUp.Api.Tests.Integration.V1.Identity
 {
-    public class IdentityControllerTests : IClassFixture<IdentityApiFactory>
+    public class IdentityControllerTests_Refresh : IClassFixture<IdentityApiFactory>
     {
         private IdentityFaker _faker;
         private readonly IdentityApiFactory _apiFactory;
         private readonly HttpClient _apiClient;
 
-        public IdentityControllerTests(IdentityApiFactory identityApiFactory)
+        public IdentityControllerTests_Refresh(IdentityApiFactory identityApiFactory)
         {
             _faker = new IdentityFaker();
             _apiFactory = identityApiFactory;
             _apiClient = identityApiFactory.CreateClient();
-        }
-
-        [Fact]
-        public async Task Register_WhenGivenValidRequest_ReturnsTokenPair()
-        {
-            var result = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Register, _faker.RegisterUserRequest);
-            result.Should().HaveStatusCode(HttpStatusCode.OK);
-
-            var response = await result.Content.ReadFromJsonAsync<IdentityResponse>();
-            response!.Token.Should().NotBeEmpty();
-            response.RefreshToken.Should().NotBeEmpty();
-        }
-
-        [Fact]
-        public async Task Register_WhenGivenInvalidRequest_ReturnsBadRequestWithValidationErrors()
-        {
-            var invalidRequest = new RegisterUserRequest()
-            {
-                Username = "BAD-",
-                Password = "i'm-G00d-and-$tr0ng-p4ssworD",
-                Device = new DeviceRequest() { Id = "i'm ok", Name = "i'm ok" }
-            };
-
-            var result = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Register, invalidRequest);
-            result.Should().HaveStatusCode(HttpStatusCode.BadRequest);
-
-            var response = await result.Content.ReadFromJsonAsync<ErrorResponse>();
-            var expectedErrors = new[] 
-            {
-                $"'{nameof(RegisterUserRequest.Username)}' may only contain lowercase letters or hyphens.",
-                $"'{nameof(RegisterUserRequest.Username)}' cannot end with a hyphen."
-            };
-            response!.Errors.Should().BeEquivalentTo(expectedErrors);
-        }
-
-        [Fact]
-        public async Task Register_WhenGivenDuplicateUsername_ReturnsBadRequestWithDetailedError()
-        {
-            var otherFaker = new IdentityFaker();
-            var duplicatedUsernameRegisterRequest = new RegisterUserRequest()
-            {
-                Username = _faker.RegisterUserRequest.Username,
-                Password = otherFaker.RegisterUserRequest.Password,
-                Device = otherFaker.RegisterUserRequest.Device,
-            };
-
-            var nonDuplicatedResult = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Register, _faker.RegisterUserRequest);
-            nonDuplicatedResult.Should().HaveStatusCode(HttpStatusCode.OK);
-
-            var duplicatedResult = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Register, duplicatedUsernameRegisterRequest);
-            duplicatedResult.Should().HaveStatusCode(HttpStatusCode.BadRequest);
-
-            var response = await duplicatedResult.Content.ReadFromJsonAsync<ErrorResponse>();
-            response!.Errors.Should().ContainSingle(e => e == "User with this username already exist.");
-        }
-
-        [Fact]
-        public async Task Login_WhenGivenValidCredentialsOfAnExistingUser_ReturnsTokenPair()
-        {
-            var resultReg = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Register, _faker.RegisterUserRequest);
-
-            var result = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Login, _faker.LoginUserRequest);
-            result.Should().HaveStatusCode(HttpStatusCode.OK);
-
-            var response = await result.Content.ReadFromJsonAsync<IdentityResponse>();
-            response!.Token.Should().NotBeEmpty();
-            response.RefreshToken.Should().NotBeEmpty();
-        }
-
-        [Fact]
-        public async Task Login_WhenGivenInvalidCredentials_ReturnsBadRequestWithGenericMessage()
-        {
-            // try to login without registering
-            var result = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Login, _faker.LoginUserRequest);
-            result.Should().HaveStatusCode(HttpStatusCode.BadRequest);
-
-            var response = await result.Content.ReadFromJsonAsync<ErrorResponse>();
-            response!.Errors.Should().ContainSingle(e => e == "Invalid request.");
         }
 
         [Fact]
@@ -117,11 +39,11 @@ namespace DevUp.Api.Tests.Integration.V1.Identity
             await Task.Delay(_apiFactory.AuthenticationOptions.TokenExpiry);
 
             // refresh and grab new token pair
-            var refreshRequest = new RefreshUserRequest() 
-            { 
-                Token = oldTokenPair!.Token!, 
-                RefreshToken = oldTokenPair.RefreshToken!, 
-                Device = _faker.LoginUserRequest.Device 
+            var refreshRequest = new RefreshUserRequest()
+            {
+                Token = oldTokenPair!.Token!,
+                RefreshToken = oldTokenPair.RefreshToken!,
+                Device = _faker.LoginUserRequest.Device
             };
             var refreshResult = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Refresh, refreshRequest);
             refreshResult.Should().HaveStatusCode(HttpStatusCode.OK);
@@ -209,11 +131,11 @@ namespace DevUp.Api.Tests.Integration.V1.Identity
             await Task.Delay(delayMs);
 
             // try refreshing
-            var refreshRequest = new RefreshUserRequest() 
-            { 
-                Token = oldTokenPair!.Token!, 
-                RefreshToken = oldTokenPair.RefreshToken!, 
-                Device = _faker.LoginUserRequest.Device 
+            var refreshRequest = new RefreshUserRequest()
+            {
+                Token = oldTokenPair!.Token!,
+                RefreshToken = oldTokenPair.RefreshToken!,
+                Device = _faker.LoginUserRequest.Device
             };
             var refreshResult = await _apiClient.PostAsJsonAsync(Route.Api.V1.Identity.Refresh, refreshRequest);
             refreshResult.Should().HaveStatusCode(HttpStatusCode.BadRequest);
