+++ backend/DevUp/tests/DevUp.Domain.Tests.Unit/Identity/Services/TokenServiceTests.cs
@@ -6,9 +6,7 @@ using DevUp.Domain.Common.Types;
 using DevUp.Domain.Identity.Entities;
 using DevUp.Domain.Identity.Repositories;
 using DevUp.Domain.Identity.Services;
-using DevUp.Domain.Identity.Services.Exceptions;
 using DevUp.Domain.Identity.Setup;
-using DevUp.Domain.Identity.ValueObjects;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using Moq;
